make sure we can pass tests with pytest
