+++ README.rst
@@ -36,6 +36,7 @@ TO DO
     
 
 - |ss| See if we can make `conditional_moments.py` inherit from base model to avoid redundancy |se|
+- Make the tests more compact for conditional moments (use the pytest arguments)
 - Test adding tf function on the run gradient descent
 - Timing of batch handler and data handler
